Introduced MapGenerator Class and Extended Grid Functionality
This commit implements an abstract MapGenerator class that forms the basis of the map generation algorithm, introducing properties for Width and Height, and the method InitializeGrid().
The commit also includes an initial implementation of LifeMapGenerator which is a concrete subclass of MapGenerator, using game of life-like algorithms to dynamically generate map layouts. It adds StartingDensity property and methods for generating starting points and different patterns.
Moreover, new functionality has been added to GeneratorGrid including a Here direction to the Direction enumeration, allowing grid cells to populate themselves, and making IsPositionSafe() method public for wider access.
In addition, refactoring was performed with the deletion of Generator class, and the property GeneratorAlgorithm was added to LifeGeneratedMap class.